requirements.txt
aiofiles==24.1.0
python-dotenv>=1.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.1.31
choicelib==0.1.5
colorama==0.4.6
frozenlist==1.5.0
idna==3.10
msgspec==0.19.0
multidict==6.3.2
propcache==0.3.1
pydantic==2.11.2
pydantic_core==2.33.1
typing-inspection==0.4.0
typing_extensions==4.13.1
vbml==1.1.post1
vkbottle==4.4.5
vkbottle-types==5.199.63.4
yarl==1.18.3
yt-dlp==2025.3.31
